Job and Salary entities are added

+++ src/main/java/com/ercanbeyen/springbootfirstrestapi/repository/EmployeeRepository.java
@@ -3,7 +3,7 @@ package com.ercanbeyen.springbootfirstrestapi.repository;
 import com.ercanbeyen.springbootfirstrestapi.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Long> { // JpaRepository<Used class, Primary key>
+public interface EmployeeRepository extends JpaRepository<Employee, Integer> { // JpaRepository<Used class, Primary key>
     /*
     User findByFirstName(String firstname);
     User findByLastName(String lastname);
